Add: Chapter 10-6 (#772)
